pyproject.toml
[tool.poetry]
name = "xaler-app"
version = "0.1.0"
description = ""
authors = ["Aleksey Maryaskin <maryaskin.a.v@gmail.com>"]
readme = "README.md"
packages = [{ include = "xaler_app" }]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
djangorestframework = "^3.14.0"
markdown = "^3.5"
django-filter = "^23.3"
drf-spectacular = "^0.26.5"
djangorestframework-simplejwt = { extras = ["crypto"], version = "^5.3.0" }
environs = "^9.5.0"
gunicorn = "^21.2.0"
django-cors-headers = "^4.3.0"
psycopg2-binary = "^2.9.9"
pillow = "^10.1.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
flake8 = "^6.1.0"
ruff = "^0.1.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    ".pyc"
]

line-length = 120
indent-width = 4

target-version = "py311"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = []

fixable = ["ALL"]

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "lf"